+++ source/main/cardano/setup.js
@@ -2,7 +2,7 @@
 import { createWriteStream, readFileSync } from 'fs';
 import { spawn, exec } from 'child_process';
 import { BrowserWindow } from 'electron';
-import { Logger } from '../../common/logging';
+import { Logger } from '../utils/logging';
 import { prepareArgs } from './config';
 import { CardanoNode } from './CardanoNode';
 import {
@@ -17,9 +17,9 @@ import { safeExitWithCode } from '../utils/safeExitWithCode';
 import type {
   TlsConfig,
   CardanoNodeState,
-  CardanoStatus,
-} from '../../common/types/cardanoNode.types';
-import { CardanoNodeStates } from '../../common/types/cardanoNode.types';
+  CardanoStatus
+} from '../../common/types/cardano-node.types';
+import { CardanoNodeStates } from '../../common/types/cardano-node.types';
 import type { LauncherConfig } from '../config';
 import {
   NODE_KILL_TIMEOUT,