+++ source/main/index.js
@@ -1,20 +1,21 @@
 // @flow
 import os from 'os';
-import { app, BrowserWindow, globalShortcut, Menu, dialog } from 'electron';
+import { app, BrowserWindow, globalShortcut, Menu, dialog, shell } from 'electron';
 import { client } from 'electron-connect';
 import { includes } from 'lodash';
-import { Logger } from '../common/logging';
+import { Logger } from './utils/logging';
 import { setupLogging } from './utils/setupLogging';
-import { makeEnvironmentGlobal } from './utils/makeEnvironmentGlobal';
 import HandleDiskSpace from './utils/HandleDiskSpace';
 import { createMainWindow } from './windows/main';
 import { winLinuxMenu } from './menus/win-linux';
 import { osxMenu } from './menus/osx';
 import { installChromeExtensions } from './utils/installChromeExtensions';
-import environment from '../common/environment';
-import { OPEN_ABOUT_DIALOG_CHANNEL } from '../common/ipc/open-about-dialog';
-import { GO_TO_ADA_REDEMPTION_SCREEN_CHANNEL } from '../common/ipc/go-to-ada-redemption-screen';
-import { GO_TO_NETWORK_STATUS_SCREEN_CHANNEL } from '../common/ipc/go-to-network-status-screen';
+import { environment } from './environment';
+import {
+  OPEN_ABOUT_DIALOG_CHANNEL,
+  GO_TO_ADA_REDEMPTION_SCREEN_CHANNEL,
+  GO_TO_NETWORK_STATUS_SCREEN_CHANNEL
+} from '../common/ipc/api';
 import mainErrorHandler from './utils/mainErrorHandler';
 import { launcherConfig } from './config';
 import { setupCardano, restartCardanoNode } from './cardano/setup';
@@ -22,7 +23,7 @@ import { CardanoNode } from './cardano/CardanoNode';
 import { safeExitWithCode } from './utils/safeExitWithCode';
 import { ensureXDGDataIsSet } from './cardano/config';
 import { acquireDaedalusInstanceLock } from './utils/lockFiles';
-import { CardanoNodeStates } from '../common/types/cardanoNode.types';
+import { CardanoNodeStates } from '../common/types/cardano-node.types';
 import type { CheckDiskSpaceResponse } from './utils/HandleDiskSpace';
 
 // Global references to windows to prevent them from being garbage collected
@@ -55,6 +56,7 @@ const restartWithoutSafeMode = async () => {
   safeExitWithCode(22);
 };
 
+const { isDev, isMacOS, isWatchMode, buildLabel } = environment;
 const menuActions = {
   openAbout,
   goToAdaRedemption,
@@ -115,24 +117,23 @@ app.on('ready', async () => {
 
   Logger.info(`========== Daedalus is starting at ${new Date().toString()} ==========`);
 
-  Logger.debug(`!!! ${environment.getBuildLabel()} is running on ${os.platform()} version ${os.release()}
+  Logger.debug(`!!! ${buildLabel} is running on ${os.platform()} version ${os.release()}
             with CPU: ${JSON.stringify(os.cpus(), null, 2)} with
             ${JSON.stringify(os.totalmem(), null, 2)} total RAM !!!`);
 
   ensureXDGDataIsSet();
-  makeEnvironmentGlobal(process.env);
-  await installChromeExtensions(environment.isDev());
+  await installChromeExtensions(isDev);
 
   cardanoNode = setupCardano(launcherConfig, mainWindow);
 
-  if (environment.isWatchMode()) {
+  if (isWatchMode) {
     // Connect to electron-connect server which restarts / reloads windows on file changes
     client.create(mainWindow);
   }
 
   // Build app menus
   let menu;
-  if (process.platform === 'darwin') {
+  if (isMacOS) {
     menu = Menu.buildFromTemplate(osxMenu(app, mainWindow, menuActions, isInSafeMode));
     Menu.setApplicationMenu(menu);
   } else {
@@ -141,7 +142,7 @@ app.on('ready', async () => {
   }
 
   // Hide application window on Cmd+H hotkey (OSX only!)
-  if (process.platform === 'darwin') {
+  if (isMacOS) {
     app.on('activate', () => {
       if (!mainWindow.isVisible()) app.show();
     });
@@ -159,6 +160,18 @@ app.on('ready', async () => {
     Logger.info('mainWindow received <close> event. Safe exiting Daedalus now.');
     event.preventDefault();
     await safeExit();
+  });
+
+  // Security feature: Prevent creation of new browser windows
+  // https://github.com/electron/electron/blob/master/docs/tutorial/security.md#14-disable-or-limit-creation-of-new-windows
+  app.on('web-contents-created', (_, contents) => {
+    contents.on('new-window', (event, url) => {
+      // Prevent creation of new BrowserWindows via links / window.open
+      event.preventDefault();
+      Logger.info(`Prevented creation of new browser window with url ${url}`);
+      // Open these links with the default browser
+      shell.openExternal(url);
+    });
   });
 
   // Wait for controlled cardano-node shutdown before quitting the app