+++ source/renderer/app/api/common/types.js
@@ -23,8 +23,19 @@ export type Pagination = {
 };
 
 export type SendBugReportRequest = {
-  email: string,
-  subject: string,
-  problem: string,
-  compressedLogsFile: string,
+  requestFormData: {
+    email: string,
+    subject: string,
+    problem: string,
+    compressedLogsFile: string,
+  },
+  environmentData: {
+    network: string,
+    version: string,
+    os: string,
+    apiVersion: string,
+    build: string,
+    installerVersion: string,
+    reportUrl: string,
+  }
 };