+++ source/renderer/app/api/utils/request.js
@@ -1,5 +1,4 @@
 // @flow
-import https from 'https';
 import { size, has, get, omit } from 'lodash';
 import querystring from 'querystring';
 import { encryptPassphrase, getContentLength } from './';
@@ -66,7 +65,7 @@ function typedRequest<Response>(
       };
     }
 
-    const httpsRequest = https.request(options);
+    const httpsRequest = global.https.request(options);
     if (hasRequestBody) {
       httpsRequest.write(requestBody);
     }