+++ source/renderer/app/containers/settings/categories/SupportSettingsPage.js
@@ -1,13 +1,11 @@
 // @flow
 import React, { Component } from 'react';
 import { observer, inject } from 'mobx-react';
-import { remote } from 'electron';
 import SupportSettings from '../../../components/settings/categories/SupportSettings';
 import type { InjectedProps } from '../../../types/injectedPropsType';
 import BugReportDialog from '../../../components/profile/bug-report/BugReportDialog';
-import { generateFileNameWithTimestamp } from '../../../../../common/fileName';
-
-const shell = require('electron').shell;
+import { generateFileNameWithTimestamp } from '../../../../../common/utils/files';
+import { openExternalUrlChannel } from '../../../ipc/open-external-url';
 
 @inject('stores', 'actions') @observer
 export default class SupportSettingsPage extends Component<InjectedProps> {
@@ -16,7 +14,7 @@ export default class SupportSettingsPage extends Component<InjectedProps> {
 
   handleExternalLinkClick = (event: MouseEvent, url: string) => {
     event.preventDefault();
-    shell.openExternal(url);
+    openExternalUrlChannel.send(url);
   };
 
   handleSupportRequestClick = () => {
@@ -26,8 +24,9 @@ export default class SupportSettingsPage extends Component<InjectedProps> {
   };
 
   handleDownloadLogs = () => {
+    // TODO: refactor this direct access to the dialog api
     const fileName = generateFileNameWithTimestamp();
-    const destination = remote.dialog.showSaveDialog({
+    const destination = global.dialog.showSaveDialog({
       defaultPath: fileName,
     });
     if (destination) {