+++ source/renderer/app/ipc/cardano.ipc.js
@@ -3,15 +3,12 @@ import type {
   CardanoNodeState, CardanoStatus,
   FaultInjectionIpcRequest,
   TlsConfig
-} from '../../../common/types/cardanoNode.types';
+} from '../../../common/types/cardano-node.types';
 import { RendererIpcChannel } from './lib/RendererIpcChannel';
 import {
-  CARDANO_AWAIT_UPDATE_CHANNEL,
-  CARDANO_STATE_CHANGE_CHANNEL,
-  CARDANO_TLS_CONFIG_CHANNEL,
-  CARDANO_RESTART_CHANNEL,
-  CARDANO_FAULT_INJECTION_CHANNEL, CARDANO_STATUS_CHANNEL
-} from '../../../common/ipc/cardano.ipc';
+  CARDANO_FAULT_INJECTION_CHANNEL, CARDANO_RESTART_CHANNEL, CARDANO_STATE_CHANGE_CHANNEL,
+  CARDANO_STATUS_CHANNEL, CARDANO_TLS_CONFIG_CHANNEL, CARDANO_AWAIT_UPDATE_CHANNEL
+} from '../../../common/ipc/api';
 
 // IpcChannel<Incoming, Outgoing>