+++ source/renderer/app/stores/NetworkStatusStore.js
@@ -13,7 +13,7 @@ import {
   SYSTEM_TIME_POLL_INTERVAL,
 } from '../config/timingConfig';
 import { UNSYNCED_BLOCKS_ALLOWED } from '../config/numbersConfig';
-import { Logger } from '../../../common/logging';
+import { Logger } from '../utils/logging';
 import { GET_DISK_SPACE_STATUS } from '../../../common/ipc-api';
 import {
   cardanoStateChangeChannel,
@@ -21,13 +21,13 @@ import {
   restartCardanoNodeChannel,
   cardanoStatusChannel,
 } from '../ipc/cardano.ipc';
-import { CardanoNodeStates } from '../../../common/types/cardanoNode.types';
+import { CardanoNodeStates } from '../../../common/types/cardano-node.types';
 import type { GetNetworkStatusResponse } from '../api/nodes/types';
 import type {
   CardanoNodeState,
   CardanoStatus,
   TlsConfig
-} from '../../../common/types/cardanoNode.types';
+} from '../../../common/types/cardano-node.types';
 import type { NodeQueryParams } from '../api/nodes/requests/getNodeInfo';
 import type { IpcEvent } from '../../../common/ipc/lib/IpcChannel';
 import type { CheckDiskSpaceResponse } from '../../../main/utils/HandleDiskSpace';