+++ source/renderer/app/stores/SidebarStore.js
@@ -1,13 +1,15 @@
 // @flow
 import { action, computed, observable } from 'mobx';
 import { get } from 'lodash';
-import { ipcRenderer } from 'electron';
 import Store from './lib/Store';
 import { sidebarConfig } from '../config/sidebarConfig';
 import { WalletSyncStateTags } from '../domains/Wallet';
-import { GO_TO_ADA_REDEMPTION_SCREEN_CHANNEL } from '../../../common/ipc/go-to-ada-redemption-screen';
+import { GO_TO_ADA_REDEMPTION_SCREEN_CHANNEL } from '../../../common/ipc/api';
 import { formattedWalletAmount } from '../utils/formatters';
 import type { SidebarWalletType } from '../types/sidebarTypes';
+
+// TODO: refactor all parts that rely on this to ipc channels!
+const { ipcRenderer } = global;
 
 export default class SidebarStore extends Store {
 
@@ -22,13 +24,17 @@ export default class SidebarStore extends Store {
     actions.toggleSubMenus.listen(this._toggleSubMenus);
     actions.activateSidebarCategory.listen(this._onActivateSidebarCategory);
     actions.walletSelected.listen(this._onWalletSelected);
+
+    // TODO: refactor to ipc channel
     ipcRenderer.on(GO_TO_ADA_REDEMPTION_SCREEN_CHANNEL, this._resetActivateSidebarCategory);
+
     this.registerReactions([
       this._syncSidebarRouteWithRouter,
     ]);
   }
 
   teardown() {
+    // TODO: refactor to ipc channel
     ipcRenderer.removeListener(
       GO_TO_ADA_REDEMPTION_SCREEN_CHANNEL,
       this._resetActivateSidebarCategory